move implementation of copy/move from app to gui
