+++ src/app/qgsguivectorlayertools.h
@@ -23,7 +23,7 @@
  * or a feature is added.
  */
 
-class QgsGuiVectorLayerTools : public QObject, public QgsVectorLayerTools
+class QgsGuiVectorLayerTools : public QgsVectorLayerTools
 {
     Q_OBJECT
 
@@ -70,19 +70,6 @@ class QgsGuiVectorLayerTools : public QObject, public QgsVectorLayerTools
      * @return            True if successful
      */
     bool saveEdits( QgsVectorLayer* layer ) const override;
-
-    /**
-     * Copy and move features with defined translation.
-     *
-     * @param layer The layer
-     * @param request The request for the features to be moved. It will be assigned to a new feature request with the newly copied features.
-     * @param dx The translation on x
-     * @param dy The translation on y
-     * @return True if all features could be copied.
-     *
-     * TODO QGIS 3: remove const qualifier
-     */
-    virtual bool copyMoveFeatures( QgsVectorLayer* layer, QgsFeatureRequest &request, double dx = 0, double dy = 0 ) const override;
 
   private:
     void commitError( QgsVectorLayer* vlayer ) const;
